Remove deploy script (#7)
The terms of the AGPL license are somewhat hard to parse when it
comes to aggregated works. We would normally put a GPL license
on the deploy script, and they probably work fine together, but
we don't really need it here either.

This commit moves the repository back to a structure in which it
really just provides PHP List for us.